requirements.txt
Django==1.11.3
djangorestframework==3.6.3
Pillow==4.3.0
